Removed `WarpEffect` class
